Modificado controlador de archivo
Fue modificado para hacer uso de la relacion y asi bloquear acciones
sobre un archivo si no es el usuario propietario del mismo.

+++ controllers/archivo.js
@@ -5,7 +5,9 @@ const fs = require('fs');
 
 // Responder con todos los Archivos
 const getArchivos = (req, res) =>{
-    Archivo.find({},(err, archivos)=>{
+    let user = req.user;
+
+    Archivo.find({ user: user },(err, archivos)=>{
         if(err) return res.status(500).send({message: `Error al realizar la peticion: ${err}`});
         if(!archivos) return res.status(404).send({message: 'No hay archivos guardados'});
         res.status(200).send(archivos)
@@ -15,10 +17,11 @@ const getArchivos = (req, res) =>{
 // Responder con INFO de un unico Archivo
 const getArchivo = (req, res) =>{
     let archivoID = req.params.fileID;
-
+    
     Archivo.findById(archivoID, (err, archivo)=>{
         if(err) return res.status(500).send({message: `Error al realizar la peticion: ${err}`});
         if(!archivo) return res.status(404).send({message: 'El archivo no existe en la Base de Datos'});
+        if(archivo.user != req.user) return res.status(403).send({message: 'No cuenta con los permisos para ver la info de este archivo'});
         res.status(200).send(archivo);
     });
 };
@@ -28,7 +31,7 @@ const buscarArchivo = (req, res) => {
     let archivoName = req.params.fileName;
     let expresion = new RegExp(archivoName,'i');
 
-    Archivo.find({'name':expresion},(err, archivos)=>{
+    Archivo.find({'name':expresion, user: req.user },(err, archivos)=>{
         if(err) return res.status(500).send({message: `Error al realizar la peticion: ${err}`});
         if(!archivos) return res.status(404).send({message: 'No se encontro ningun archivo en la Base de Datos con el nombre brindado'});
         res.status(200).send(archivos);
@@ -54,17 +57,17 @@ const subirArchivo = (req, res) =>{
     
     let archivo = new Archivo();
     let archivosSubidos = req.files.archivos;
-    let ubicacion;
-    
+    let ubicacion = './upload/' + req.user + '/' + archivosSubidos.name;
+
     archivo.name = archivosSubidos.name;
     archivo.size = archivosSubidos.size;
     archivo.type = archivosSubidos.mimeTipe;
-    ubicacion = './upload/' + archivosSubidos.name;
     archivo.location = ubicacion;
     archivosSubidos.mv(ubicacion);
     let date = new Date;
     archivo.add_Date = date.getDate() + "/" + (date.getMonth() + 1) + "/" + date.getFullYear();
     archivo.add_Date += " " + date.getHours() + ":" + date.getMinutes();
+    archivo.user = req.user;
 
     archivo.save((err, archivoGuardado)=>{
         if(err) res.status(500).send({message: `Error al guardar en la DB: ${err}`});
@@ -90,13 +93,18 @@ const borrarArchivo = (req, res)=>{
     Archivo.findById(archivoID, (err, archivo)=>{
         if(err) return res.status(500).send({message: `Hubo un error intentar borrar del archivo: ${err}`});
         if(!archivo) return res.status(404).send({message: 'No se encontro el archivo'});
-        archivo.remove(err => {
-            if(err) return res.status(500).send({message: `Hubo un error intentar borrar del archivo en la DB: ${err}`});
-            fs.unlink(archivo.location, (err)=>{
-                if(err) return res.status(500).send({message:`Hubo un error al intentar Borrar el archivo en el servidor`});
-                res.status(200).send(archivo);
+        if(archivo.user = req.user){
+            archivo.remove(err => {
+                if(err) return res.status(500).send({message: `Hubo un error intentar borrar del archivo en la DB: ${err}`});
+                fs.unlink(archivo.location, (err)=>{
+                    if(err) return res.status(500).send({message:`Hubo un error al intentar Borrar el archivo en el servidor`});
+                    res.status(200).send(archivo);
+                })
             })
-        })
+        }
+        else{
+            return res.status(403).send({message: 'No tiene permisos para eliminar este archivo'});
+        }
     })
 }
